% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.h.R
\name{single}
\alias{single}
\title{Single Variable}
\usage{
single(data, variable)
}
\arguments{
\item{data}{the data as a data frame, with variables encoded as numeric
(for interval/ratio variables), ordered (for ordinal variables), or factor
(for nominal variables)}

\item{variable}{a string naming a variable from \code{data}}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$advice} \tab \tab \tab \tab \tab the method recommendation \cr
  \code{results$plot} \tab \tab \tab \tab \tab a scatter plot of the data \cr
}
}
\description{
Provides a recommended statistical method given the measurement level of 
your data. Appropriate when you are interested in a single variable.
}
\examples{
exampleData <- data.frame(x1 = rnorm(20),
                          x2 = as.factor(c(rep(1, 10), rep(2, 10))),
                          x3 = as.ordered(c(rep(1, 10), rep(2, 5), rep(3, 5))))

single(exampleData, variable = 'x1')
single(exampleData, variable = 'x2')
single(exampleData, variable = 'x3')

}
