% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlational.h.R
\name{correlational}
\alias{correlational}
\title{Correlational Research}
\usage{
correlational(data, dep, independents, controls)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming variable 1 / the dependent variable from
\code{data}}

\item{independents}{a string or vector of strings naming variable 2 / the
independent variable(s) from \code{data}}

\item{controls}{a string or vector of strings naming the control
variable(s) from \code{data}}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$advice} \tab \tab \tab \tab \tab a html \cr
}
}
\description{
Find an appropriate method given the measurement level of your data - 
correlational research
}
\examples{
exampleData <- data.frame(x1 = rnorm(20),
                          x2 = as.factor(c(rep(1, 10), rep(2, 10))),
                          x3 = as.ordered(rep(1, 5), rep(2, 5), rep(3, 10)),
                          x4 = rnorm(20),
                          x5 = rnorm(20))

correlational(exampleData, dep = 'x3', independents = 'x2')
correlational(exampleData, dep = 'x1', independents = c('x2', 'x3'), controls = c('x4', 'x5'))

}
