% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated.h.R
\name{repeated}
\alias{repeated}
\title{Repeated Measurements}
\usage{
repeated(data, variables)
}
\arguments{
\item{data}{the data as a data frame, with variables encoded as numeric
(for interval/ratio variables), ordered (for ordinal variables), or factor
(for nominal variables)}

\item{variables}{a vector of strings naming the related variables from
\code{data}}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$advice} \tab \tab \tab \tab \tab the method recommendation \cr
  \code{results$plot} \tab \tab \tab \tab \tab a scatter plot of the data \cr
}
}
\description{
Provides a recommended statistical method given the measurement level of 
your data. Appropriate for repeated measurements.
}
\examples{
exampleData <- data.frame(x1 = rnorm(20),
                          x2 = rnorm(20),
                          x3 = rnorm(20),
                          x4 = as.factor(c(rep(1, 10), rep(2, 10))),
                          x5 = as.factor(c(rep(1, 5), rep(2, 5), rep(3, 10))),
                          x6 = as.ordered(c(rep(1, 5), rep(2, 5), rep(3, 10))),
                          x7 = as.ordered(c(rep(1, 10), rep(2, 5), rep(3, 5))))

repeated(exampleData, variables = c('x1', 'x2', 'x3'))
repeated(exampleData, variables = c('x4', 'x5'))
repeated(exampleData, variables = c('x6', 'x7'))

}
