% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.h.R
\name{single}
\alias{single}
\title{Single Variable}
\usage{
single(data, variable)
}
\arguments{
\item{data}{the data as a data frame}

\item{variable}{a string naming a variable from \code{data}}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$advice} \tab \tab \tab \tab \tab a html \cr
}
}
\description{
Find an appropriate method given the measurement level of your data - 
single variable
}
\examples{
exampleData <- data.frame(x1 = rnorm(20),
                          x2 = as.factor(c(rep(1, 10), rep(2, 10))),
                          x3 = as.ordered(rep(1, 10), rep(2, 5), rep(3, 5)))

single(exampleData, variable = 'x1')
single(exampleData, variable = 'x2')
single(exampleData, variable = 'x3')

}
